+++ django_storage_url/__init__.py
@@ -1,6 +1,8 @@
 from .backends import get_storage_class, register_storage_class
 from .base import (
-    dsn_configured_storage, dsn_configured_storage_class, get_storage,
+    dsn_configured_storage,
+    dsn_configured_storage_class,
+    get_storage,
 )