+++ django_storage_url/backends/__init__.py
@@ -1,5 +1,6 @@
-from django.core.files.storage import \
-    get_storage_class as django_get_storage_class
+from django.core.files.storage import (
+    get_storage_class as django_get_storage_class,
+)
 
 
 SCHEMES = {